% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgat_function.R
\name{filter_N}
\alias{filter_N}
\title{Filter sequences by N content}
\usage{
filter_N(fastafile, N_filter)
}
\arguments{
\item{fastafile}{List. A named list of DNA sequences (e.g. from
\code{seqinr::read.fasta()}).}

\item{N_filter}{Integer. Maximum number of N bases allowed in a sequence.
Sequences with more N's than this threshold will be removed.}
}
\value{
List. Filtered list of sequences.
}
\description{
Filters a list of DNA sequences by removing those with too many ambiguous
(N) bases.
}
\details{
This function is useful for quality control before phylogenetic analysis.
Sequences with excessive ambiguous bases can affect the accuracy of distance
calculations and tree construction.
}
\examples{
test_seqs <- list(
    good_seq = "ATCGATCG",
    bad_seq  = "ATCGNNNNNATCG",
    okay_seq = "ATCGNNATCG"
)
filtered <- filter_N(test_seqs, N_filter = 3)
length(filtered)  # Should be 2

}
