% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgat_function.R
\name{fastafile_new}
\alias{fastafile_new}
\title{Filter FASTA sequences by N content}
\usage{
fastafile_new(fastafile, N_filter)
}
\arguments{
\item{fastafile}{List. A list of DNA sequences read by seqinr::read.fasta()}

\item{N_filter}{Integer. Maximum number of N bases allowed in a sequence.
Sequences with more N's than this threshold will be removed.}
}
\value{
List. Filtered FASTA sequences.
}
\description{
Filters a FASTA file by removing sequences with too many ambiguous (N) bases.
}
\details{
This function is useful for quality control before phylogenetic analysis.
Sequences with excessive ambiguous bases can affect the accuracy of distance
calculations and tree construction.
}
\examples{
\dontrun{
library(seqinr)
fasta <- read.fasta("sequences.fasta", seqtype = "DNA", as.string = TRUE)
filtered_fasta <- fastafile_new(fasta, N_filter = 50)
}

}
